Fixit some problems and add exceptions and validation

+++ user-service/src/main/java/com/example/userservice/service/UserServiceImplementation.java
@@ -4,6 +4,8 @@ import com.example.userservice.dto.UserCreateDTO;
 import com.example.userservice.dto.UserDetailDTO;
 import com.example.userservice.dto.UserSummaryDTO;
 import com.example.userservice.entity.User;
+import com.example.userservice.exception.UserAlreadyExistsException;
+import com.example.userservice.exception.UserNotFoundException;
 import com.example.userservice.mapper.UserMapper;
 import com.example.userservice.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -20,23 +22,35 @@ public class UserServiceImplementation implements UserService {
 
     @Override
     public UserDetailDTO createUser(UserCreateDTO userDto) {
+        if(userRepository.existsByEmail(userDto.getEmail())) {
+            throw new UserAlreadyExistsException(userDto.getEmail());
+        }
         return userMapper.toDetailDto(userRepository.save(userMapper.toEntity(userDto)));
     }
 
     @Override
     public UserDetailDTO getUserById(Long id) {
-        return userMapper.toDetailDto(userRepository.findById(id).orElseThrow(RuntimeException::new));
+        return userMapper.toDetailDto(userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id)));
     }
 
     @Override
     public UserDetailDTO updateUser(Long id, UserCreateDTO userDto) {
-        User user = userMapper.toEntity(userDto);
-        user.setId(id);
-        return userMapper.toDetailDto(userRepository.save(user));
+     //   User user = userMapper.toEntity(userDto);// ты здесь создаешь нового пользователя а не получаешь по id
+        User existingUser = userRepository.findById(id)
+                .orElseThrow(() -> new UserNotFoundException(id));
+
+        existingUser.setName(userDto.getName());
+        existingUser.setEmail(userDto.getEmail());
+
+        User updatedUser = userRepository.save(existingUser);
+        return userMapper.toDetailDto(updatedUser);
     }
 
     @Override
     public void deleteUser(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
     }
